+++ src/AliasUnit.cpp
@@ -254,7 +254,7 @@ void AliasUnit::removeAlias( TAlias * pT )
 }
 
 
-qint64 AliasUnit::getNewID()
+int AliasUnit::getNewID()
 {
     return ++mMaxID;
 }
